+++ com.unity.formats.alembic/Editor/Importer/AlembicImporterEditor.cs
@@ -419,8 +419,8 @@ namespace UnityEditor.Formats.Alembic.Importer
                 GUILayout.FlexibleSpace();
 
 #if !HAIR_AVAILABLE
-                string msg = L10n.Tr("There is no Hair package installed. " +
-                                     "Install the hair package first to generate " +
+                string msg = L10n.Tr("Hair package not found. " +
+                                     "You have to install it first to generate " +
                                      "a curve-based groom.");
 
                 GUILayout.FlexibleSpace();
@@ -464,8 +464,9 @@ namespace UnityEditor.Formats.Alembic.Importer
                     GUILayout.Button(L10n.Tr("Generate Hair Asset"));
                     GUI.enabled = true;
 
-                    var message = L10n.Tr("Unable to locate curves in the alembic asset. " +
-                                          "Ensure that the asset contains curves and \"Import Curves\" is enabled.");
+                    var message = L10n.Tr("Unable to locate curves in the Alembic asset. " +
+                                          "Ensure that the asset contains curves and \"Import Curves\" is " +
+                                          "enabled in the \"Model\" tab.");
                     EditorGUILayout.HelpBox(message, MessageType.Warning);